requirements.txt
discord
openai
asyncio
requests
python-dotenv